requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
spacy>=3.7.0
python-dateutil>=2.8.0
psycopg2-binary>=2.9.0
httpx>=0.27.0
numpy>=1.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
